I reinstalled the fragtag program and for some reason it's no longer working. Somehow there must have been some change in the code since it was last installed.
The code was really hard to read so I'm working
on refactoring it. It still doesn't work.
